+++ CHANGELOG.md
@@ -1,0 +1,3 @@
+### Master
+
+- Convert to TypeScript
