+++ LICENSE
@@ -1,6 +1,10 @@
 The MIT License (MIT)
 
-Copyright (c) 2018 Artsy Inc. <dev@artsymail.com>
+Copyright (c) 2018 Artsy Inc.
+
+Christopher Pappas <damassi.pappas@gmail.com>
+Alan Johnson <alan@artsymail.com>
+Artsy <dev@artsymail.com>
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
