+++ tsconfig.json
@@ -1,15 +1,11 @@
 {
   "compilerOptions": {
     "allowSyntheticDefaultImports": true,
-    "experimentalDecorators": true,
     "jsx": "react",
     "lib": ["dom", "es2015", "es2016", "es2017"],
     "module": "commonjs",
     "moduleResolution": "node",
     "target": "es2015",
-    "noImplicitAny": false,
-    "noUnusedLocals": false,
-    "noUnusedParameters": true,
     "outDir": "dist",
     "rootDir": "src",
     "declaration": true
